fix: overflow using int8_t, changed to uint16_t

+++ Core/Inc/Drivers/sg90_driver.h
@@ -23,6 +23,6 @@
  * @param channel Timer channel associated to the PWM.
  * @param degrees Position servo in degrees.
  */
-void SG90Driver_SetPosition(TIM_HandleTypeDef* htim, uint32_t channel, int8_t degrees);
+void SG90Driver_SetPosition(TIM_HandleTypeDef* htim, uint32_t channel, int16_t degrees);
 
 #endif /* __SG90_DRIVER_H */
